Moved Meshes and terrainHeightmap to assets

+++ fruithunter/Source/Drawable/Mesh/Mesh.cpp
@@ -188,7 +188,7 @@ void Mesh::createBuffers(bool instancing) {
 	}
 }
 void Mesh::bindMaterial(int index) { m_materials[index].bind(); }
-int Mesh::getVertexCount() const { return m_meshVertices.size(); }
+int Mesh::getVertexCount() const { return (int)m_meshVertices.size(); }
 const std::vector<Vertex>& Mesh::getVertexPoints() const { return m_meshVertices; }
 const Microsoft::WRL::ComPtr<ID3D11Buffer> Mesh::getVertexBuffer() const { return m_vertexBuffer; }
 std::string Mesh::getName() const { return m_loadedMeshName; }
